+++ controllers/src/test/scala/cromwell/pipeline/controller/AuthControllerTest.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.model.ContentTypes.`application/json`
 import akka.http.scaladsl.model.{ HttpEntity, StatusCodes }
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import cromwell.pipeline.controller.AuthController._
-import cromwell.pipeline.datastorage.dto.formatters.AuthFormatters.{ AuthResponse, PasswordProblemsResponse, SignInRequest, SignUpRequest }
+import cromwell.pipeline.datastorage.dto.auth.{ AuthResponse, PasswordProblemsResponse, SignInRequest, SignUpRequest }
 import cromwell.pipeline.datastorage.utils.validator.DomainValidation
 import cromwell.pipeline.service.AuthService
 import org.scalamock.scalatest.MockFactory
