+++ portal/src/it/scala/cromwell/pipeline/controller/UserControllerItTest.scala
@@ -7,7 +7,8 @@ import com.typesafe.config.Config
 import cromwell.pipeline.ApplicationComponents
 import cromwell.pipeline.datastorage.dao.repository.utils.TestUserUtils
 import cromwell.pipeline.datastorage.dto.User
-import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.{ PasswordUpdateRequest, UserNoCredentials, UserUpdateRequest }
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters._
+import cromwell.pipeline.datastorage.dto.{ PasswordUpdateRequest, UserNoCredentials, UserUpdateRequest }
 import cromwell.pipeline.datastorage.utils.auth.AccessTokenContent
 import cromwell.pipeline.utils.TestContainersUtils
 import de.heikoseeberger.akkahttpplayjson.PlayJsonSupport
