+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/User.scala
@@ -1,7 +1,7 @@
 package cromwell.pipeline.datastorage.dto
 
 import cromwell.pipeline.datastorage.dto.User.UserEmail
-import cromwell.pipeline.datastorage.dto.formatters.UserFormatters.ProfilePicture
+import cromwell.pipeline.datastorage.dto.formatters.UserFormatters._
 import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
 import play.api.libs.json.{ Format, Json, OFormat }
@@ -27,4 +27,29 @@ final case class UserId(value: String) extends MappedTo[String]
 
 object UserId {
   implicit lazy val userIdFormat: Format[UserId] = implicitly[Format[String]].inmap(UserId.apply, _.value)
+}
+
+final case class ProfilePicture(value: Array[Byte]) extends MappedTo[Array[Byte]]
+
+final case class PasswordUpdateRequest(currentPassword: String, newPassword: String, repeatPassword: String)
+
+final case class UserUpdateRequest(email: String, firstName: String, lastName: String)
+
+final case class UserNoCredentials(
+  userId: UserId,
+  email: String,
+  firstName: String,
+  lastName: String,
+  active: Boolean
+)
+
+object UserNoCredentials {
+  def fromUser(user: User): UserNoCredentials =
+    UserNoCredentials(
+      userId = user.userId,
+      email = user.email,
+      firstName = user.firstName,
+      lastName = user.lastName,
+      active = user.active
+    )
 }
