+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/AuthResponse.scala
@@ -1,0 +1,3 @@
+package cromwell.pipeline.datastorage.dto.auth
+
+final case class AuthResponse(accessToken: String, refreshToken: String, accessTokenExpiration: Long)
