+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/PasswordProblemsResponse.scala
@@ -1,0 +1,3 @@
+package cromwell.pipeline.datastorage.dto.auth
+
+final case class PasswordProblemsResponse(value: String, valid: Boolean = false, errors: List[Map[String, String]])
