+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/SignInRequest.scala
@@ -1,0 +1,3 @@
+package cromwell.pipeline.datastorage.dto.auth
+
+final case class SignInRequest(email: String, password: String)
