+++ repositories/src/main/scala/cromwell/pipeline/datastorage/dto/auth/SignUpRequest.scala
@@ -1,0 +1,3 @@
+package cromwell.pipeline.datastorage.dto.auth
+
+final case class SignUpRequest(email: String, password: String, firstName: String, lastName: String)
