+++ services/src/main/scala/cromwell/pipeline/service/ProjectVersioning.scala
@@ -4,7 +4,7 @@ import java.nio.file.Path
 
 import cromwell.pipeline.datastorage.dto.{ Project, ProjectFile, Version }
 import cromwell.pipeline.datastorage.dto.Project
-import cromwell.pipeline.datastorage.dto.formatters.ProjectFormatters.{ ProjectFile, Version }
+import cromwell.pipeline.datastorage.dto.{ ProjectFile, Version }
 
 import scala.concurrent.Future
